refactor: split core crud into submodules
