+++ pkgs/standards/autoapi/autoapi/v3/core/crud/helpers.py
@@ -1,11 +1,10 @@
-# autoapi/v3/core/crud.py
+# autoapi/v3/core/crud/helpers.py
 from __future__ import annotations
 
 from typing import (
     Any,
     Dict,
     Iterable,
-    List,
     Mapping,
     Optional,
     Sequence,
@@ -516,343 +515,3 @@ def _normalize_list_call(
         "db": db,
         "sort": sort,
     }
-
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Canonical CRUD
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-async def create(
-    model: type, data: Mapping[str, Any], db: Union[Session, AsyncSession]
-) -> Any:
-    """
-    Insert a single row. Returns the persisted model instance.
-    Flush-only (commit happens later in END_TX).
-    """
-    data = _filter_in_values(model, data or {}, "create")
-    _validate_enum_values(model, data)
-    obj = model(**data)
-    if hasattr(db, "add"):
-        db.add(obj)
-    await _maybe_flush(db)
-    return obj
-
-
-async def read(model: type, ident: Any, db: Union[Session, AsyncSession]) -> Any:
-    """
-    Load a single row by primary key. Raises NoResultFound if not found.
-    """
-    obj = await _maybe_get(db, model, ident)
-    if obj is None:
-        raise NoResultFound(f"{model.__name__}({ident!r}) not found")
-    return obj
-
-
-async def update(
-    model: type, ident: Any, data: Mapping[str, Any], db: Union[Session, AsyncSession]
-) -> Any:
-    """
-    Partial update by primary key. Missing keys are left unchanged.
-    Returns the updated model instance. Flush-only.
-    """
-    data = _filter_in_values(model, data or {}, "update")
-    _validate_enum_values(model, data)
-    obj = await read(model, ident, db)
-    skip = _immutable_columns(model, "update")
-    _set_attrs(obj, data, allow_missing=True, skip=skip)
-    await _maybe_flush(db)
-    return obj
-
-
-async def replace(
-    model: type, ident: Any, data: Mapping[str, Any], db: Union[Session, AsyncSession]
-) -> Any:
-    """
-    PUT semantics with upsert behaviour.
-
-    If the row exists it is replaced entirely (missing attributes are nulled).
-    If the row does not exist it is created with the provided identifier.
-    Flush-only.
-    """
-    data = _filter_in_values(model, data or {}, "replace")
-    _validate_enum_values(model, data)
-    pk = _single_pk_name(model)
-    obj = await _maybe_get(db, model, ident)
-    if obj is None:
-        payload = {pk: ident, **data}
-        return await create(model, payload, db=db)
-    skip = _immutable_columns(model, "replace")
-    _set_attrs(obj, data, allow_missing=False, skip=skip)
-    await _maybe_flush(db)
-    return obj
-
-
-async def merge(
-    model: type, ident: Any, data: Mapping[str, Any], db: Union[Session, AsyncSession]
-) -> Any:
-    """PATCH semantics with upsert behaviour."""
-    pk = _single_pk_name(model)
-    ident = _coerce_pk_value(model, ident)
-    obj = await _maybe_get(db, model, ident)
-
-    # Respect create-only fields when upserting a new row
-    verb = "update" if obj is not None else "create"
-    data = _filter_in_values(model, data or {}, verb)
-    _validate_enum_values(model, data)
-    data_no_pk = {k: v for k, v in data.items() if k != pk}
-    if obj is None:
-        payload = {pk: ident, **data_no_pk}
-        return await create(model, payload, db=db)
-    skip = _immutable_columns(model, "update")
-    _set_attrs(obj, data_no_pk, allow_missing=True, skip=skip)
-    await _maybe_flush(db)
-    return obj
-
-
-async def delete(
-    model: type, ident: Any, db: Union[Session, AsyncSession]
-) -> Dict[str, int]:
-    """
-    Delete by primary key. Returns {"deleted": 1} if removed, else raises NoResultFound.
-    Flush-only.
-    """
-    obj = await read(model, ident, db)
-    await _maybe_delete(db, obj)
-    await _maybe_flush(db)
-    return {"deleted": 1}
-
-
-# NOTE: tolerant signature: accepts positional/keyword and ignores stray args
-async def list(*_args: Any, **_kwargs: Any) -> List[Any]:  # noqa: A001  (shadow built-in)
-    """
-    Simple list with equality filters + skip/limit (+ optional sort).
-    Tolerant to:
-      - missing filters (defaults to {})
-      - accidental bound-method 'self' (first positional arg)
-      - positional or keyword args
-      - stray extras (e.g., request) which are ignored
-    """
-    model, params = _normalize_list_call(_args, _kwargs)
-
-    filters: Mapping[str, Any] = _coerce_filters(model, params["filters"])
-    skip: Optional[int] = params["skip"]
-    limit: Optional[int] = params["limit"]
-    db: Union[Session, AsyncSession] = params["db"]
-    sort = params["sort"]
-
-    if select is None:  # pragma: no cover
-        # Fallback: legacy query API
-        q = db.query(model)  # type: ignore[attr-defined]
-        if filters:
-            q = q.filter_by(**filters)  # type: ignore[attr-defined]
-        if isinstance(skip, int):
-            q = q.offset(max(skip, 0))  # type: ignore[attr-defined]
-        if isinstance(limit, int) and limit is not None:
-            q = q.limit(max(limit, 0))  # type: ignore[attr-defined]
-        return _builtins.list(q.all())  # type: ignore[attr-defined]
-
-    where = _apply_filters(model, filters)
-    stmt = select(model)
-    if where is not None:
-        stmt = stmt.where(where)
-
-    order_exprs = _apply_sort(model, sort)
-    if order_exprs:
-        for ob in order_exprs:
-            stmt = stmt.order_by(ob)
-
-    if isinstance(skip, int):
-        stmt = stmt.offset(max(skip, 0))
-    if isinstance(limit, int) and limit is not None:
-        stmt = stmt.limit(max(limit, 0))
-
-    result = await _maybe_execute(db, stmt)
-    return _builtins.list(result.scalars().all())  # type: ignore[attr-defined]
-
-
-async def clear(
-    *args: Any,
-    **kwargs: Any,
-) -> Dict[str, int]:
-    """
-    Delete many rows matching equality filters. Returns {"deleted": N}.
-    Flush-only. Tolerant to the same calling variations as `list`.
-    """
-    # Reuse normalizer to accept the same shapes
-    model, params = _normalize_list_call(args, kwargs)
-    raw_filters: Mapping[str, Any] = params["filters"]
-    db: Union[Session, AsyncSession] = params["db"]
-
-    if sa_delete is None:  # pragma: no cover
-        # Fallback path: manual iteration
-        items = await list(model, raw_filters, db=db)
-        n = 0
-        for obj in items:
-            await _maybe_delete(db, obj)
-            n += 1
-        await _maybe_flush(db)
-        return {"deleted": n}
-
-    filt = _coerce_filters(model, raw_filters)
-    where = _apply_filters(model, filt)
-    stmt = sa_delete(model)
-    if where is not None:
-        stmt = stmt.where(where)
-
-    res = await _maybe_execute(db, stmt)
-    await _maybe_flush(db)
-    # Some dialects don't populate rowcount; best-effort
-    n = int(getattr(res, "rowcount", 0) or 0)
-    return {"deleted": n}
-
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Bulk variants
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-async def bulk_create(
-    model: type, rows: Iterable[Mapping[str, Any]], db: Union[Session, AsyncSession]
-) -> List[Any]:
-    """
-    Insert many rows. Returns the list of persisted instances.
-    Flush-only.
-    """
-    items_data = [dict(r) for r in (rows or ())]
-    for r in items_data:
-        _validate_enum_values(model, r)
-    items = [model(**r) for r in items_data]
-    if not items:
-        return []
-    if hasattr(db, "add_all"):
-        db.add_all(items)  # type: ignore[attr-defined]
-    else:
-        for it in items:
-            db.add(it)  # type: ignore[attr-defined]
-    await _maybe_flush(db)
-    return items
-
-
-async def bulk_update(
-    model: type, rows: Iterable[Mapping[str, Any]], db: Union[Session, AsyncSession]
-) -> List[Any]:
-    """
-    Update many rows by PK. Each row must include the PK field.
-    Returns the list of updated instances. Flush-only.
-    """
-    pk = _single_pk_name(model)
-    skip = _immutable_columns(model, "update")
-    updated: List[Any] = []
-    for r in rows or ():
-        r = dict(r)
-        _validate_enum_values(model, r)
-        ident = r.get(pk)
-        if ident is None:
-            raise ValueError(f"bulk_update requires '{pk}' in each row")
-        obj = await read(model, ident, db)
-        # remove pk to avoid accidental overwrite
-        data = {k: v for k, v in r.items() if k != pk}
-        _set_attrs(obj, data, allow_missing=True, skip=skip)
-        updated.append(obj)
-    if updated:
-        await _maybe_flush(db)
-    return updated
-
-
-async def bulk_replace(
-    model: type, rows: Iterable[Mapping[str, Any]], db: Union[Session, AsyncSession]
-) -> List[Any]:
-    """
-    Replace many rows by PK. Each row must include the PK field.
-    Missing attributes are nulled (except PK). Flush-only.
-    """
-    pk = _single_pk_name(model)
-    skip = _immutable_columns(model, "replace")
-    replaced: List[Any] = []
-    for r in rows or ():
-        r = dict(r)
-        _validate_enum_values(model, r)
-        ident = r.get(pk)
-        if ident is None:
-            raise ValueError(f"bulk_replace requires '{pk}' in each row")
-        obj = await read(model, ident, db)
-        data = {k: v for k, v in r.items() if k != pk}
-        _set_attrs(obj, data, allow_missing=False, skip=skip)
-        replaced.append(obj)
-    if replaced:
-        await _maybe_flush(db)
-    return replaced
-
-
-async def bulk_merge(
-    model: type, rows: Iterable[Mapping[str, Any]], db: Union[Session, AsyncSession]
-) -> List[Any]:
-    """Merge many rows by primary key with upsert semantics."""
-    pk = _single_pk_name(model)
-    results: List[Any] = []
-    to_create: List[Mapping[str, Any]] = []
-    for r in rows or ():
-        r = dict(r)
-        ident = _coerce_pk_value(model, r.get(pk))
-        if ident is not None:
-            existing = await _maybe_get(db, model, ident)
-            if existing is not None:
-                data = {k: v for k, v in r.items() if k != pk}
-                merged = await merge(model, ident, data, db=db)
-                results.append(merged)
-                continue
-            r[pk] = ident
-        to_create.append(r)
-    if to_create:
-        created = await bulk_create(model, to_create, db)
-        results.extend(created)
-    return results
-
-
-async def bulk_delete(
-    model: type, idents: Iterable[Any], db: Union[Session, AsyncSession]
-) -> Dict[str, int]:
-    """
-    Delete many rows by a sequence of PK values. Returns {"deleted": N}.
-    Flush-only.
-    """
-    pk_name = _single_pk_name(model)
-    id_seq = _builtins.list(idents or ())
-    if not id_seq:
-        return {"deleted": 0}
-
-    # Prefer DELETE ... WHERE pk IN (...)
-    if sa_delete is not None:
-        col = getattr(model, pk_name)
-        stmt = sa_delete(model).where(col.in_(id_seq))  # type: ignore[attr-defined]
-        res = await _maybe_execute(db, stmt)
-        await _maybe_flush(db)
-        n = int(getattr(res, "rowcount", 0) or 0)
-        return {"deleted": n}
-
-    # Fallback: row-by-row
-    n = 0
-    for ident in id_seq:
-        obj = await read(model, ident, db)
-        await _maybe_delete(db, obj)
-        n += 1
-    await _maybe_flush(db)
-    return {"deleted": n}
-
-
-__all__ = [
-    "create",
-    "read",
-    "update",
-    "replace",
-    "merge",
-    "delete",
-    "list",
-    "clear",
-    "bulk_create",
-    "bulk_update",
-    "bulk_replace",
-    "bulk_merge",
-    "bulk_delete",
-]
